+++ src/test/java/com/google/cloud/spanner/jdbc/JdbcParameterStoreTest.java
@@ -126,10 +126,10 @@ public class JdbcParameterStoreTest {
     assertEquals(1, ((Long) params.getParameter(1)).longValue());
     verifyParameter(params, Value.int64(1));
     params.setParameter(1, (float) 1, Types.FLOAT);
-    assertEquals(1.0f, ((Float) params.getParameter(1)).floatValue(), 0.0f);
+    assertEquals(1.0f, (Float) params.getParameter(1), 0.0f);
     verifyParameter(params, Value.float64(1));
     params.setParameter(1, (double) 1, Types.DOUBLE);
-    assertEquals(1.0d, ((Double) params.getParameter(1)).doubleValue(), 0.0d);
+    assertEquals(1.0d, (Double) params.getParameter(1), 0.0d);
     verifyParameter(params, Value.float64(1));
     params.setParameter(1, new Date(1970 - 1900, 0, 1), Types.DATE);
     assertEquals(new Date(1970 - 1900, 0, 1), params.getParameter(1));
@@ -204,10 +204,10 @@ public class JdbcParameterStoreTest {
       assertEquals(1, ((Long) params.getParameter(1)).longValue());
       verifyParameter(params, Value.int64(1));
       params.setParameter(1, (float) 1, type);
-      assertEquals(1.0f, ((Float) params.getParameter(1)).floatValue(), 0.0f);
+      assertEquals(1.0f, (Float) params.getParameter(1), 0.0f);
       verifyParameter(params, Value.int64(1));
       params.setParameter(1, (double) 1, type);
-      assertEquals(1.0d, ((Double) params.getParameter(1)).doubleValue(), 0.0d);
+      assertEquals(1.0d, (Double) params.getParameter(1), 0.0d);
       verifyParameter(params, Value.int64(1));
       params.setParameter(1, BigDecimal.ONE, type);
       assertEquals(BigDecimal.ONE, params.getParameter(1));
@@ -229,10 +229,10 @@ public class JdbcParameterStoreTest {
       assertEquals(1, ((Long) params.getParameter(1)).longValue());
       verifyParameter(params, Value.float64(1));
       params.setParameter(1, (float) 1, type);
-      assertEquals(1.0f, ((Float) params.getParameter(1)).floatValue(), 0.0f);
+      assertEquals(1.0f, (Float) params.getParameter(1), 0.0f);
       verifyParameter(params, Value.float64(1));
       params.setParameter(1, (double) 1, type);
-      assertEquals(1.0d, ((Double) params.getParameter(1)).doubleValue(), 0.0d);
+      assertEquals(1.0d, (Double) params.getParameter(1), 0.0d);
       verifyParameter(params, Value.float64(1));
       params.setParameter(1, BigDecimal.ONE, type);
       assertEquals(BigDecimal.ONE, params.getParameter(1));
@@ -328,10 +328,10 @@ public class JdbcParameterStoreTest {
       assertEquals(1, ((Long) params.getParameter(1)).longValue());
       verifyParameter(params, Value.bool(true));
       params.setParameter(1, (float) 1, type);
-      assertEquals(1.0f, ((Float) params.getParameter(1)).floatValue(), 0.0f);
+      assertEquals(1.0f, (Float) params.getParameter(1), 0.0f);
       verifyParameter(params, Value.bool(true));
       params.setParameter(1, (double) 1, type);
-      assertEquals(1.0d, ((Double) params.getParameter(1)).doubleValue(), 0.0d);
+      assertEquals(1.0d, (Double) params.getParameter(1), 0.0d);
       verifyParameter(params, Value.bool(true));
       params.setParameter(1, BigDecimal.ZERO, type);
       assertEquals(BigDecimal.ZERO, params.getParameter(1));
@@ -357,10 +357,10 @@ public class JdbcParameterStoreTest {
       assertEquals(1, ((Long) params.getParameter(1)).longValue());
       verifyParameter(params, Value.numeric(BigDecimal.ONE));
       params.setParameter(1, (float) 1, type);
-      assertEquals(1.0f, ((Float) params.getParameter(1)).floatValue(), 0.0f);
+      assertEquals(1.0f, (Float) params.getParameter(1), 0.0f);
       verifyParameter(params, Value.numeric(BigDecimal.valueOf(1.0)));
       params.setParameter(1, (double) 1, type);
-      assertEquals(1.0d, ((Double) params.getParameter(1)).doubleValue(), 0.0d);
+      assertEquals(1.0d, (Double) params.getParameter(1), 0.0d);
       verifyParameter(params, Value.numeric(BigDecimal.valueOf(1.0)));
     }
   }
@@ -498,10 +498,10 @@ public class JdbcParameterStoreTest {
     assertEquals(1, ((Long) params.getParameter(1)).longValue());
     verifyParameter(params, Value.int64(1));
     params.setParameter(1, (float) 1, (Integer) null);
-    assertEquals(1.0f, ((Float) params.getParameter(1)).floatValue(), 0.0f);
+    assertEquals(1.0f, (Float) params.getParameter(1), 0.0f);
     verifyParameter(params, Value.float64(1));
     params.setParameter(1, (double) 1, (Integer) null);
-    assertEquals(1.0d, ((Double) params.getParameter(1)).doubleValue(), 0.0d);
+    assertEquals(1.0d, (Double) params.getParameter(1), 0.0d);
     verifyParameter(params, Value.float64(1));
     params.setParameter(1, new Date(1970 - 1900, 0, 1), (Integer) null);
     assertEquals(new Date(1970 - 1900, 0, 1), params.getParameter(1));
